+++ z/util/regex/rgxscan.cpp
@@ -20,7 +20,7 @@ namespace z
 			int amount = 0;
 			bool canGreedy = false;
 
-			for (size_t i=0; i<pattern.length(); ++i)
+			for (int i=0; i<pattern.length(); ++i)
 			{
 				auto ch = pattern[i];
