+++ z/util/regex/rules/character.cpp
@@ -6,7 +6,7 @@ namespace z
 	{
 		namespace rgx
 		{
-			character::character(uint32_t value, bool insensitive, size_t min, size_t max, bool greedy):
+			character::character(uint32_t value, bool insensitive, int min, int max, bool greedy):
 				rule(min,max,greedy), value(value), insensitive(insensitive) {}
 
 			bool character::match(uint32_t current) const
